test(popover): migrate inline
